Add signal handling and a graceful exit to slurpyd
Killing slurpyd with SIGINT or SIGTERM now goes through a graceful exit
method which removes any files in the process of being written or
compressed. This uses global arrays WRITING_FILES and COMPRESSING_FILES
to check for any files in this stage.

It will be sensible in the future to implement a decorator which times
writing functions and populates these arrays without constantly having
to add the code manually.