Add read_config() method to slurpyd
This method streamlines the reading process and removes the need for the
user to call os.path.expanduser() each time a writable path is desired.
In such a case, the user should use 'output_sh' for log config, or
'out_dir_sh' for other configs.

Implementing this also required the test_slurpyd suite to be rewritten
to load test-specific config files through the new read_config() method.
Doing so automatically set the appropriate _sh fields to their expanded
forms, which enable compatibility with the slurpyd methods.